LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/run/caretnavigation-expected.txt

line 1

line 2

line 2

FAIL Move with ArrowLeft to BR following a div assert_equals: expected (string) "div" but got (undefined) undefined
FAIL Move with ArrowLeft to BR following an img assert_equals: expected (string) "div" but got (undefined) undefined
